Add files via upload
Problem Statement

The problem of health insurance prediction aims to build a model that can accurately predict the cost of health insurance for an individual based on various demographic and health-related factors. This information is critical for insurance providers to set fair and affordable premiums for their customers and for individuals to make informed decisions about their health coverage. The model should take into account factors such as age, gender, location, smoking habits, and pre-existing medical conditions to make a prediction. The challenge is to create a model that accurately balances the trade-off between risk and cost and provides reliable and fair predictions for a diverse range of individuals. The solution to this problem will have significant implications for the health insurance industry, helping providers make better business decisions and ensuring that individuals have access to affordable health coverage.